--- .idea/material_theme_project_new.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="MaterialThemeProjectNewConfig">
-    <option name="metadata">
-      <MTProjectMetadataState>
-        <option name="migrated" value="true" />
-        <option name="pristineConfig" value="false" />
-        <option name="userId" value="-270b5ce4:191086168e8:-7ffe" />
-      </MTProjectMetadataState>
-    </option>
-  </component>
-</project>
