+++ chat/models/message_request.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class MessageRequest(BaseModel):
+    message: str
+    user_id: int
