+++ product_info/models/product.py
@@ -1,0 +1,1 @@
+import
